Fix llava_llama_3_8b_v1_1_gguf: explicitly cast model to target dtype to fix float16/bfloat16 mismatch

+++ llava_llama_3_8b_v1_1_gguf/conditional_generation/pytorch/loader.py
@@ -96,7 +96,7 @@ class ModelLoader(ForgeModel):
             model = LlavaForConditionalGeneration(config)
         finally:
             torch.set_default_dtype(old_default_dtype)
-        model.eval()
+        model.to(target_dtype).eval()
 
         self.config = model.config
         return model
